pyproject.toml
[project]
name = "pdf-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-storage-queue>=12.12.0",
    "python-dotenv>=1.1.0",
]
